refactor: split md catch-all into individual route files
