+++ src/app/md/[place]/loadouts/[loadout]/route.ts
@@ -1,0 +1,11 @@
+import { createMarkdownRouteHandler } from '@/server/utils/createMarkdownRoute';
+
+const handler = createMarkdownRouteHandler();
+
+export async function GET(
+  request: Request,
+  { params }: { params: Promise<{ loadout: string; place: string }> },
+) {
+  const { loadout, place } = await params;
+  return handler(request, `/md/${place}/loadouts/${loadout}`);
+}
