+++ src/app/md/[place]/shells.md/route.ts
@@ -1,0 +1,57 @@
+import { markdownTable } from 'markdown-table';
+
+import { createCache } from '@/server/utils/createCache';
+import {
+  escapeMarkdownLink,
+  formatMarkdown,
+} from '@/server/utils/formatMarkdown';
+import {
+  getCachedMarkdown,
+  mdResponse,
+  resolvePlaceFromParams,
+} from '@/server/utils/mdResponse';
+import { getPlaceFromName } from '@/utils/placeUtils';
+import { getConfig } from '@generated/config';
+import { getShells } from '@generated/shells';
+
+import type { PlaceName } from '@generated/config';
+
+async function revalidate(placeName: PlaceName) {
+  const shells = getShells();
+  const { data: config } = getConfig();
+  const place = getPlaceFromName(config, placeName);
+  const shellsData = shells.data[place.placeId]?.data;
+  if (!shellsData) return null;
+
+  const weaponTables = Object.entries(shellsData).map(
+    ([weapon, shellsList]) => {
+      return `## ${weapon}\n\n${markdownTable([
+        ['Shell name', 'Type'],
+        ...shellsList.map((shell) => [
+          `[${escapeMarkdownLink(shell.name)}](/md/${place.initials}/shells/${shell.slug}.md)`,
+          shell.type,
+        ]),
+      ])}`;
+    },
+  );
+  return formatMarkdown(`# Shells\n\n${weaponTables.join('\n\n')}`);
+}
+
+const cache = createCache<PlaceName, string | null>(revalidate);
+
+export async function GET(
+  _request: Request,
+  { params }: { params: Promise<{ place: string }> },
+) {
+  const { place: initials } = await params;
+  const { baseUrl, placeName } = resolvePlaceFromParams(initials);
+  if (!placeName) return new Response(null, { status: 404 });
+
+  const markdown = await getCachedMarkdown(cache, revalidate, placeName);
+  if (!markdown) return new Response(null, { status: 404 });
+
+  return mdResponse(
+    markdown,
+    new URL(`${initials}/shells`, baseUrl).toString(),
+  );
+}
