+++ src/app/md/[place]/teams.md/route.ts
@@ -1,0 +1,59 @@
+import { markdownTable } from 'markdown-table';
+
+import { createCache } from '@/server/utils/createCache';
+import {
+  escapeMarkdownLink,
+  formatMarkdown,
+} from '@/server/utils/formatMarkdown';
+import {
+  getCachedMarkdown,
+  mdResponse,
+  resolvePlaceFromParams,
+} from '@/server/utils/mdResponse';
+import { getPlaceFromName } from '@/utils/placeUtils';
+import { getConfig } from '@generated/config';
+import { getLoadouts } from '@generated/loadouts';
+
+import type { PlaceName } from '@generated/config';
+
+async function revalidate(placeName: PlaceName) {
+  const loadouts = getLoadouts();
+  const { data: config } = getConfig();
+  const place = getPlaceFromName(config, placeName);
+  const loadoutsData = loadouts.data[place.placeId];
+  if (!loadoutsData) return null;
+
+  const teams = loadoutsData.metadata.teams;
+  const table = markdownTable([
+    ['Team name', 'Loadouts'],
+    ...teams.map((team) => {
+      const teamLoadouts = Object.entries(loadoutsData.data)
+        .filter(([, loadout]) => loadout.teams.includes(team))
+        .map(([name]) => name);
+      return [
+        `[${escapeMarkdownLink(team)}](/md/${place.initials}/teams/${encodeURIComponent(team)}.md)`,
+        teamLoadouts.join(', '),
+      ];
+    }),
+  ]);
+  return formatMarkdown(`# Teams\n\n${table}`);
+}
+
+const cache = createCache<PlaceName, string | null>(revalidate);
+
+export async function GET(
+  _request: Request,
+  { params }: { params: Promise<{ place: string }> },
+) {
+  const { place: initials } = await params;
+  const { baseUrl, placeName } = resolvePlaceFromParams(initials);
+  if (!placeName) return new Response(null, { status: 404 });
+
+  const markdown = await getCachedMarkdown(cache, revalidate, placeName);
+  if (!markdown) return new Response(null, { status: 404 });
+
+  return mdResponse(
+    markdown,
+    new URL(`${initials}/teams`, baseUrl).toString(),
+  );
+}
